ERPExplorer.create_plot! — Method

create_plot!(plots, data, vars, scatter_styles, line_styles, continuous_vars)

Arguments:
- plots - a SpecApi list to push into.
- data - a DataFrame to be subsetted.
- vars contains the levels to be plotted.
Return Value: .

ERPExplorer.extract_variables — Method

extract_variables(model)

Takes the Unfold model and extract variables from it for future functions.

Arguments:
- model::UnfoldLinearModel{Float64} - Unfold linear model with categorical and continuous terms.

These variables are:

names - all terms of the model.

symbols- non-numeric terms of the model.

types - types of terms.

vals - min, max, variance, mean for continuous terms, levels for categorical terms.

Return Values: Vector{Pair{Symbol}}.

ERPExplorer.formular_widgets — Method

formular_widgets(variables)

Creates widgets to control each variable of a model.
Arguments:
- variables::Vector{Pair{Symbol}} - vector of key-value pairs with information about the model formula terms.

Actions:

Extract ranges of term values.

Create widgets for each term (Slider for continuous, SelectSet for categorical).

Creating the formula with checkboxes and translating to HTML code.

Convert checkboxes to Observables.

Return Values:

widget_checkbox: Dictionary with the current values of the widgets (term => values).

widget_signal: widget_checkbox but as Observable, a signal that emits a dictionary with the current values of the widgets.

formular_widget: The HTML element that can be displayed to interact with the the widgets.

value_ranges: A dictionaryy containing the value ranges of each formula term.

ERPExplorer.mapping_dropdowns — Method

mapping_dropdowns(varnames, var_types)

Maps dropdown menus on the left panel of the Figure.
Arguments:
- varnames::Vector{Symbol} - vector of the model formula terms.

var_types::Vector{Symbol} - vector of types of the model formula terms.

Actions:

Take categorical variables and put their values into dropdown menus.

There will be 5 menus for: color, markers, line styles, column and row facets.

Map each menu object with its name on the Figure.

Create HTML containers using Document Object Model (DOM) from Bonito.

Arrange containers on the panel using Col() and Row(). Specify their styling.

Return Values:

mapping::Observable{Dict{Symbol, Symbol}} - interactive dictionary with menus and their default value.

mapping_dom::Hyperscript.Node{Hyperscript.HTMLSVG} - dropdown menus in HTML code with styling and layout.

ERPExplorer.plot_data — Method

plot_data(data, value_ranges, categorical_vars, continuous_vars, mapping_obs)

data: effects(Dict(...), m)

value_ranges:

categorical_vars:

continuous_vars:

mapping: Dict name=>property.

Return Value: DataFrames.

Internally, we use a PlotConfig struct to keep track of common plotting options, so that all functions have a similar API.