ERPExplorer.create_plot! — Method

create_plot!(plots, data, vars, scatter_styles, line_styles, continuous_vars)

Arguments:
- plots - a SpecApi list to push into.
- data - a dataframe to be subsetted.
- vars contains the levels to be plotted.
Return Value: .

ERPExplorer.formular_widgets — Method

formular_widgets(variables)

Creates widgets to control each variable of a model.
Return Values:
* widget_signal: a signal that emits a dictionary with the current values of the widgets.

formular_widget: The HTML element that can be displayed to interact with the the widgets

value_ranges: A dictionary with the value ranges of each variable.

ERPExplorer.plot_data — Method

plot_data(data, value_ranges, categorical_vars, continuous_vars, mapping_obs)
- `data`: effects(Dict(...), m) 
- `value_ranges`:
- `categorical_vars`:
- `continuous_vars`:
- `mapping`: Dict name=>property

Return Value: DataFrames.

Internally, we use a PlotConfig struct to keep track of common plotting options, so that all functions have a similar API.