ERPExplorer.define_scatter_line_style! — Method

define_scatter_line_style!(plots, data, vars, scatter_styles, line_styles, continuous_vars)

Define styling of lines and points (scatter).

subset the data.

select points and plot scatter. Define scatter style: markersize and color.

plot lines and define line style: colormap, color range, color.

Arguments:
- plots::Vector{Makie.PlotSpec} - an empty SpecApi list to push into parts of the layout.
- data::DataFrame - a DataFrame with predicted values to be subsetted.
- vars::Dict{Any, Any} contains the levels to be plotted.
- scatter_styles::Dict{Any, Any} - define colors of scatter.

line_styles:: Dict{Symbol, Pair{Symbol, Tuple{Tuple{String, String}, Symbol}}} - define line styles: colormap, color range, color.

continuous_vars::Vector{Symbol} - continuous terms.

Return Value: Makie.GridLayoutSpec.

ERPExplorer.explore — Method

explore(model::UnfoldModel; positions = nothing, size = (700, 600))

Run the dashboard for explorative ERP analysis.

Arguments:
- model::UnfoldLinearModel{Float64} - Unfold linear model with categorical and continuous terms.

positions::Vector{Point{2, Float32}} - x an y coordinates of the channels.

size::Tuple{Float64, Float64} - size of the topoplot panel.

Actions:

Extract formula terms and itheir features.

Create interactive formula with checkboxes.

Arrange and map dropdown menus.

Create interactive topoplot.

Create Observable DataFrame with predicted values (yhats) and more.

Create GridLayout.

Use Base.ReentrantLock, a synchronization primitive_ to manage concurrent access to shared resources in multi-threaded programs

Create Figure.

Translate into into HTML code using DOMs.

Return Value: Hyperscript.Node{Hyperscript.HTMLSVG} - final HTML code of the dashboard.

ERPExplorer.extract_variables — Method

extract_variables(model)

Takes the Unfold model and extract variables from it for future functions.

Arguments:
- model::UnfoldLinearModel{Float64} - Unfold linear model with categorical and continuous terms.

These variables are:

names - all terms of the model.

symbols- non-numeric terms of the model.

types - types of terms.

vals - min, max, variance, mean for continuous terms, levels for categorical terms.

Return Values: Vector{Pair{Symbol}}.

ERPExplorer.formular_widgets — Method

formular_widgets(variables)

Creates widgets to control each variable of a model.
Arguments:
- variables::Vector{Pair{Symbol}} - vector of key-value pairs with information about the model formula terms.

Actions:

Extract ranges of term values.

Create widgets for each term (Slider for continuous, SelectSet for categorical).

Creating the formula with checkboxes and translating to HTML code.

Convert checkboxes to Observables.

Return Values:

widget_checkbox: Dictionary with the current values of the widgets (term => values).

widget_signal: widget_checkbox but as Observable, a signal that emits a dictionary with the current values of the widgets.

formular_widget: The HTML element that can be displayed to interact with the the widgets.

value_ranges: A dictionary containing the value ranges of each formula term.

ERPExplorer.mapping_dropdowns — Method

mapping_dropdowns(varnames, var_types)

Map and arrange dropdown menus on the left panel of the dashboard.
Arguments:
- varnames::Vector{Symbol} - vector of the model formula terms.

var_types::Vector{Symbol} - vector of types of the model formula terms.

Actions:

Take categorical variables and put their values into dropdown menus.

There will be 5 menus for: color, markers, line styles, column and row facets.

Map each menu object with its name on the Figure.

Create HTML containers using Document Object Model (DOM) from Bonito.

Arrange containers on the panel using Col() and Row(). Specify their styling.

Return Values:

mapping::Observable{Dict{Symbol, Symbol}} - interactive dictionary with menus and their default value.

mapping_dom::Hyperscript.Node{Hyperscript.HTMLSVG} - dropdown menus in HTML code with styling and layout.

ERPExplorer.plot_data — Method

plot_data(data, value_ranges, categorical_vars, continuous_vars, mapping_obs)

Plotting an interactive dashboard.

Arguments:

data::DataFrame - the result of effects(Dict(...), model) with columns: yhat, channel, dummy, time, eventname and unique columns for each formula term..

value_ranges::Vector{Pair{Symbol}} - value range for continuous variables, levels for categorical.

categorical_vars::Vector{Symbol} - categorical terms.

continuous_vars::Vector{Symbol} - continuous terms.

mapping::Dict{Symbol, Symbol} - dictionary with dropdown menus and their default values.

Action:

Create default palettes for colors, markers, line styles, and color styles for continuous values.

Check that the terms are not empty.

Plot the dashboard.

Define line and scatter styles for the line plot.

Add line and scatter styles to the legend.

Return Value: Makie.GridLayoutSpec.

ERPExplorer.topoplot_widget — Method

topoplot_widget(positions; size = ())

Controls the topoplot in the lower left panel of the figure.
Highlight the location of the current electrode and allows electrode selection.

Arguments:
- positions::Vector{Point{2, Float32}} - x an y coordinates of the channels.

size::Tuple{Float64, Float64} - size of the topoplot panel.

Actions:

Create interactive scatter.

Highlight the selected electrode with white color, others are grayed out.

Create a topolot with a null interpolator and define its style and behavior.

Hide decorations and spines.

Return Values:

h_topo::Makie.FigureAxisPlot - topoplot widget.

interactive_scatter::Observable{Int64} - number of the selected channel.

ERPExplorer.yhats_signal — Method

yhats_signal(model, widget_signal, channel)

Creates a dictionary with yhat values and more.
Arguments:
- model::UnfoldLinearModel{Float64} - vector of key-value pairs with information about the model formula terms.

widget_signal::Observable{Vector{Any}} - a signal that emits a dictionary with the current values of the widgets.

channel::Observable{Int64} - number of selected channel-

Actions:

Compute predicted value (yhat) of the given model using effects.

Create DataFrame with columns: yhat, channel, dummy, time, eventname and unique columns for each formula term.

Make it Observable.

Return Value: yhats_signal::Observable{Any} containing DataFrame with yhats.

Internally, we use a PlotConfig struct to keep track of common plotting options, so that all functions have a similar API.